Restructure enumerator to allow different MDES versions. #2574.
