+++ lib/ncs_navigator/staff_portal/warehouse/two_point_zero.rb
@@ -1,0 +1,6 @@
+require 'ncs_navigator/staff_portal'
+module NcsNavigator::StaffPortal::Warehouse
+  module TwoPointZero
+    autoload :Enumerator, 'ncs_navigator/staff_portal/warehouse/two_point_zero/enumerator'
+  end
+end
